requirements.txt
aiofiles==23.2.1
aiogram==3.9.0
aiogram-dialog==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.3
certifi==2024.7.4
environs==11.0.0
frozenlist==1.4.1
h11==0.14.0
html5lib==1.1
idna==3.7
Jinja2==3.1.4
lxml==5.2.2
magic-filter==1.0.12
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyPDF2==3.0.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
aiofiles==23.2.1
aiogram==3.9.0
aiogram-dialog==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
DataRecorder==3.5.3
DownloadKit==2.0.4
DrissionPage==4.0.5.6
environs==11.0.0
et-xmlfile==1.1.0
filelock==3.16.0
frozenlist==1.4.1
h11==0.14.0
html5lib==1.1
idna==3.7
Jinja2==3.1.4
lxml==5.2.2
magic-filter==1.0.12
MarkupSafe==2.1.5
marshmallow==3.21.3
MouseInfo==0.1.3
multidict==6.0.5
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
psutil==6.0.0
PyAutoGUI==0.9.54
pydantic==2.8.2
pydantic_core==2.20.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
requests==2.32.3
requests-file==2.1.0
rubicon-objc==0.4.9
selenium==4.9.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tldextract==5.1.2
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.2.2
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
wsproto==1.2.0
yarl==1.9.4
keyboard==0.13.5
sqlalchemy==2.0.35
yookassa==3.3.0